# tauri/packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Springtale management API",
    "description": "Local-first bot automation. Binds loopback; bearer-authenticated.",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "GET /events — paginated event log.",
        "description": "Returns recent events (trigger type, connector, timestamp, action taken).\nEvent payloads are NOT stored (ephemeral in PipelineContext per privacy model).\nThe limit clamp lives in `operations::events::list`.",
        "operationId": "events_list",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "connector",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of the event log",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}